refactor: system-monitoring to server monitoring
